Add database search
Currently, it is difficult to find a particular database for dictionary
sources with a large number of databases (such as the default dict.org
source), since there is no search function. This commit adds a search
bar to the database chooser which filters the displayed databases based
on the search query.